change the amount parameter to float64

+++ src/maney_test.go
@@ -42,9 +42,9 @@ type Money struct {
 }
 
 func (m Money) Times(multiplier int) Money {
-	return Money{amount: m.amount * multiplier, currency: m.currency}
+	return Money{amount: m.amount * float64(multiplier), currency: m.currency}
 }
 
 func (m Money) Divide(divisor int) Money {
-	return Money{amount: m.amount / divisor, currency: m.currency}
+	return Money{amount: m.amount / float64(divisor), currency: m.currency}
 }
